+++ bridge/pkg/processor/quorum_test.go
@@ -1,4 +1,4 @@
-package main
+package processor
 
 import (
 	"fmt"
